requirements.txt
pathlib
zxcvbn
python-docx
matplotlib